+++ src/api/configs-from-pkg.ts
@@ -3,10 +3,16 @@ import { BundlibPkgJson } from "./pkg";
 import pkgToConfigs from "./pkg-to-configs";
 import { BundlibRollupOptions } from "./types";
 
-async function configsFromPkg(cwd: string, dev?: boolean, pkgJson?: BundlibPkgJson): Promise<BundlibRollupOptions[]> {
+async function configsFromPkg(
+  cwd: string,
+  dev?: boolean,
+  watch?: boolean,
+  pkgJson?: BundlibPkgJson,
+): Promise<BundlibRollupOptions[]> {
   return await pkgToConfigs(
     await analizePkg(cwd, pkgJson),
     dev,
+    watch,
   );
 }